+++ packages/ui/src/components/atoms/Dropdown.tsx
@@ -1,4 +1,4 @@
-import React, {
+import {
   useState,
   useRef,
   forwardRef,
@@ -12,6 +12,10 @@ import { FlexBox, colors, dimensions, font } from '../../styles'
 import { Button } from './Button'
 import { Icon } from './Icon'
 
+const StyledFlexBox = styled(FlexBox)`
+  width: 100%;
+`
+
 const StyledDropdown = styled.div`
   cursor: pointer;
   width: 100%;
@@ -23,12 +27,16 @@ const StyledIcon = styled(Icon)`
 `
 
 const StyledImage = styled.img`
-width: 24px;
-height: 24px;
-margin-right: 10px;
+  width: 24px;
+  height: 24px;
+  margin-right: 10px;
 `
 
-const DropdownHeader = styled(Button)`
+const DeselectIcon = styled(Icon)`
+  margin-left: auto;
+`
+
+export const DropdownHeader = styled(Button)`
   justify-content: space-between;
   background-color: ${colors.white};
   padding: ${dimensions.spacing.base};
@@ -62,17 +70,30 @@ const DropdownList = styled.div`
   overflow-y: auto;
 `
 
-const DropdownItem = styled.div`
-  padding: 10px;
+const DropdownItem = styled.div<{ $isSelected: boolean }>`
+  padding: 12px 10px;
   cursor: pointer;
   font-family: ${font.fontFamily};
-  font-size: ${font.xss};
+  font-size: ${font.xs};
   display: flex;
   align-items: center;
+  color: ${({ $isSelected }) =>
+    $isSelected ? `${colors.white}` : `${colors.gray.gray3}`};
+  background-color: ${({ $isSelected }) =>
+    $isSelected ? `${colors.primaryLight}` : `${colors.white}`};
+
+  ${StyledIcon} {
+    color: ${({ $isSelected }) =>
+      $isSelected ? `${colors.white}` : `${colors.gray.gray3}`};
+  }
 
   &:hover {
     background-color: ${colors.primary};
     color: ${colors.white};
+
+    ${StyledIcon} {
+      color: ${colors.white};
+    }
   }
 `
 
@@ -86,54 +107,52 @@ export type TDropdownOption = {
 export type TDropdown = HTMLAttributes<HTMLElement> & {
   options: TDropdownOption[]
   placeholder?: string
-  defaultValue?: string
-  onValueChange?: (value: string) => void
+  defaultSelectedOption?: TDropdownOption
+  onValueChange?: (selectedOption?: TDropdownOption | undefined) => void
   openText?: string
   closeText?: string
+  deselectText?: string
+  className?: string
 }
 
 export const Dropdown = forwardRef<HTMLDivElement, TDropdown>(
   (
     {
       options = [],
-      defaultValue = '',
+      defaultSelectedOption,
       placeholder = 'Selecciona',
       onValueChange,
       openText = 'Ampliar',
       closeText = 'Cerrar',
+      deselectText = 'Borra la selección',
+      className = '',
     },
     ref
   ) => {
     const [isDropdownOpen, setIsDropdownOpen] = useState(false)
-    const [selectedValue, setSelectedValue] = useState(defaultValue)
+    const [selectedValue, setSelectedValue] = useState<
+      TDropdownOption | undefined
+    >(defaultSelectedOption)
     const dropdownListRef = useRef<HTMLDivElement>(null)
     const dropdownCloseOutsideRef = useRef<HTMLDivElement>(null)
 
     const handleSelect = useCallback(
-      (value: string) => {
-        setSelectedValue(value)
+      (selectedOption: TDropdownOption) => {
+        setSelectedValue(selectedOption)
         setIsDropdownOpen(false)
         if (onValueChange) {
-          onValueChange(value)
+          onValueChange(selectedOption)
         }
-
-        const handleClick = (event: MouseEvent) => {
-          const target = event.target as HTMLElement
-          if (dropdownListRef.current?.contains(target)) {
-            handleSelect(value);
-          }
-        }
-
-        if (dropdownListRef.current) {
-          dropdownListRef.current.addEventListener('click', handleClick)
-          return () => {
-            dropdownListRef.current?.removeEventListener('click', handleClick)
-          }
-        }
-        return () => {}
       },
-      [dropdownListRef, onValueChange]
+      [onValueChange]
     )
+
+    const handleDeselect = () => {
+      setSelectedValue(undefined)
+      if (onValueChange) {
+        onValueChange(undefined)
+      }
+    }
 
     useEffect(() => {
       const handleClickOutside = (event: MouseEvent) => {
@@ -147,30 +166,37 @@ export const Dropdown = forwardRef<HTMLDivElement, TDropdown>(
     }, [dropdownListRef, dropdownCloseOutsideRef])
 
     const selectedOption = useMemo(
-      () => options.find((option) => option.id === selectedValue),
+      () => options.find((option) => option.id === selectedValue?.id),
       [options, selectedValue]
     )
 
     return (
-      <div ref={ref}>
+      <div ref={ref} className={className}>
         <StyledDropdown data-testid="dropdown" ref={dropdownCloseOutsideRef}>
           <DropdownHeader
             data-testid="dropdown-header"
             onClick={() => setIsDropdownOpen(!isDropdownOpen)}
           >
-            { selectedOption ? ( 
-                  <FlexBox direction='row' key={selectedOption.id}>
-                  {selectedOption.icon && (
+            {selectedOption ? (
+              <StyledFlexBox
+                direction="row"
+                key={selectedOption.id}
+                justify="flex-start"
+              >
+                {selectedOption.icon && (
                   <StyledIcon name={selectedOption.icon} />
-                  )}
-                  {selectedOption.iconSvg && (
-                  <StyledImage 
-                  src={selectedOption.iconSvg} alt={selectedOption.name} />
-                  )}
-                  <span>{selectedOption.name}</span>
-                  </FlexBox>
-              ) : ( <span>{defaultValue || placeholder}</span> )}
-            
+                )}
+                {selectedOption.iconSvg && (
+                  <StyledImage
+                    src={selectedOption.iconSvg}
+                    alt={selectedOption.name}
+                  />
+                )}
+                <span>{selectedOption.name}</span>
+              </StyledFlexBox>
+            ) : (
+              <span>{placeholder}</span>
+            )}
             <StyledIcon
               name={isDropdownOpen ? 'expand_less' : 'expand_more'}
               aria-hidden="true"
@@ -179,11 +205,31 @@ export const Dropdown = forwardRef<HTMLDivElement, TDropdown>(
           </DropdownHeader>
           {isDropdownOpen && (
             <DropdownList ref={dropdownListRef}>
-              {options.map(({ name, id, icon, iconSvg }) => (
-                <DropdownItem key={id} data-testid={id} id={id} onClick={() => handleSelect(id)}>
-                  {icon && <StyledIcon name={icon} />}
-                  {iconSvg && <StyledImage src={iconSvg} alt={name} />}
-                  <span>{name}</span>
+              {options.map((option) => (
+                <DropdownItem
+                  key={option.id}
+                  data-testid={option.id}
+                  id={option.id}
+                  onClick={
+                    option.id === selectedOption?.id
+                      ? () => setIsDropdownOpen(false)
+                      : () => handleSelect(option)
+                  }
+                  $isSelected={option.id === selectedOption?.id}
+                >
+                  {option.icon && <StyledIcon name={option.icon} />}
+                  {option.iconSvg && (
+                    <StyledImage src={option.iconSvg} alt={option.name} />
+                  )}
+                  <span>{option.name}</span>
+                  {option.id === selectedOption?.id ? (
+                    <DeselectIcon
+                      name="close"
+                      onClick={() => handleDeselect()}
+                      title={deselectText}
+                      data-testid={`deselect-${option.id}`}
+                    />
+                  ) : null}
                 </DropdownItem>
               ))}
             </DropdownList>