+++ services/wiki/src/openapi/routes/resources/favorites.ts
@@ -1,12 +1,11 @@
 import { pathRoot } from '../../../routes/routes'
-import { registry } from '../../registry'
 import { resourceFavoriteSchema } from '../../../schemas'
-import { z } from '../../zod'
 import {
   invalidTokenResponse,
-  missingTokenResponse,
   userNotFoundResponse,
 } from '../../components/responses/authMiddleware'
+import { registry } from '../../registry'
+import { z } from '../../zod'
 
 registry.registerPath({
   method: 'get',
@@ -33,8 +32,7 @@ registry.registerPath({
         },
       },
     },
-    401: missingTokenResponse,
+    401: invalidTokenResponse,
     404: userNotFoundResponse,
-    498: invalidTokenResponse,
   },
 })