+++ web/usuaris/src/helpers/index.ts
@@ -1,1 +1,1 @@
-export { getUsers, loginUserFetcher, userInfoFetcher } from './fetchers'
+export { getItineraries, getUsers, loginUserFetcher } from './fetchers'